Gate 8: CLI end-to-end tests, Makefile, CI, docs
make render regenerates everything; CI lints, tests, and validates all
programs (a PR touching programs/*.json fails on invalid data). mypy
strict on the pure core, ruff clean throughout.

+++ src/towerkit/scale.py
@@ -76,4 +76,4 @@ def retention_depth(
     """Depth of a retention block as a fraction of the retention band height."""
     if max_retention <= 0 or amount <= 0:
         return 0.0
-    return (amount / max_retention) ** exponent
+    return float((amount / max_retention) ** exponent)
